==> 3_Check_by_conllu.ipynb <==
### 3. Check Entities by dependecy trees

import pandas as pd
import stanza
import re
from spacy_conll import init_parser
from tokenization import *
from keywords import keywords_dict

#### Reading files

entities = pd.read_csv('../data/entities_may.csv', index_col=[0])
# entities = entities.set_index('link', drop=False)
entities.columns

def names_and_kw_str_to_list(names_and_kw_str):
    if pd.notna(names_and_kw_str):
        res = []
        for ent in names_and_kw_str.split('<+>'):
            parts = ent.split('<#>')
            names = re.findall('(\(\d+\, \d+\))<§§>PERS<§§>(.*?)<§§>([\d\.]+)', parts[2])
            for n in names:
                r = [int(i) for i in n[0].strip('()').split(', ')]
                res.append( ( int(parts[0]), parts[1], r, n[1], float(n[2]) ) )
        return res
    return None

news = pd.read_csv('../data/may.csv', index_col=[0])
# news = news.set_index('link', drop=False)
news['filt_kw_names_str'] = entities['filt_kw_names_str']
news['kw_and_ent'] = entities.kw_and_ent
news.columns

del entities

news['filt_kw_names'] = news.filt_kw_names_str.str.split('§')
news['entities'] = news.kw_and_ent.apply(names_and_kw_str_to_list)
# news['sentenized'] = news.sentences_joined.str.split('<§>')

news['checked_with_conllu'] = None

news['all_text'] = news.title.str.cat(news.text, sep='\n', na_rep = '')
news['all_text'] = news.all_text.str.strip()

%%time
news['sentenized'] = news.apply(lambda row: tokenize_to_sent_str(row.all_text, row.language), axis=1)

#### Loading stanza models

stanza.download('uk')
stanza.download('ru', package='gsd', processors='tokenize,pos,lemma,depparse')

nlp = stanza.Pipeline('uk', processors='tokenize,lemma')

nlp_uk = init_parser(
        "stanza",
        "uk", 
        is_tokenized = True,
        include_headers=False,
        parser_opts = {'processors': 'tokenize,pos,lemma,depparse'}
    )

nlp_ru = init_parser(
        "stanza",
        "ru", 
        is_tokenized = True,
        include_headers=False,
        parser_opts = {'package': 'gsd', 'processors': 'tokenize,pos,lemma,depparse'}
    )

#### Checking sentences with keywords with dependency trees

all_keywords = '|'.join([r'\b' + kw for kw in keywords_dict])


without_check = [
    '(е|э)ксперт',
    'анал(і|и)тик',
    '(про)?цит(ує|ат|ував|увала?)',
]
according_to_keywords = [
    'за\s(словами|оцінк)',
    'на\sдумку',
    'по\sсловам',
    'по\s(оценк|мнен)',
    'згідно',
    'согласно',
    'посилання',
]
wrong_keywords = [
    'посила(вся|лася)',
]

without_check = '|'.join([r'\b' + kw for kw in without_check])
according_to_keywords = '|'.join([r'\b' + kw for kw in according_to_keywords])
wrong_keywords = '|'.join([r'\b' + kw for kw in wrong_keywords])

def check_nsubj_relation(conllu_df, ent_range, all_keywords):
    for i in ent_range:
        id = i+1
        curr_node = conllu_df.loc[id]
        if curr_node['upostag'] in ['NOUN', 'PROPN']:
            while curr_node['deprel'] != 'root':
                head_node = conllu_df.loc[curr_node['head']]
#                 print(curr_node.form, head_node.form)
                if curr_node['deprel'].startswith('nsubj'):
                    if re.match(all_keywords, head_node['form'], flags=re.I):
                        return True
                    else:
                        return False
                curr_node = head_node
    return False


def check_nmod_relation(conllu_df, ent_range, keyword):
    for i in ent_range:
        id = i+1
        curr_node = conllu_df.loc[id]
        if curr_node['upostag'] in ['NOUN', 'PROPN']:
            while curr_node['deprel'] != 'root':
                head_node = conllu_df.loc[curr_node['head']]
                if re.match(keyword, curr_node['form'], flags=re.I) or \
                    re.match(keyword, head_node['form'], flags=re.I):
                    return True
                curr_node = head_node
    return False



def sentence_check(entitites_list, sentences, filt_names_list, language):
    if isinstance(entitites_list, list) and isinstance(filt_names_list, list):
        res = []
        sent_num = -1
        for ent in entitites_list:
            if (ent[3] in filt_names_list):
                if re.search(wrong_keywords, ent[1], flags=re.I):
                    res.append(False)
                elif re.search(without_check, ent[1], flags=re.I):
                    res.append(True)
                else:
                    if sent_num != ent[0]:
                        sent_num = ent[0]
                        if language == 'uk':
                            doc = nlp_uk(sentences[sent_num])
                        else:
                            doc = nlp_ru(sentences[sent_num])
                        conllu_df = (doc._.conll_pd).set_index('id')
                    
                    if re.search(according_to_keywords, ent[1], flags=re.I):
                        keyword = ent[1].split()[-1]
                        res.append(check_nmod_relation(conllu_df, ent[2][:-1], keyword))
                    else:
                        res.append(check_nsubj_relation(conllu_df, ent[2][:-1], all_keywords))        
            else:
                res.append('NINL')
        try:
            del conllu_df
        except:
            pass
        
        return res
    return None

out_file = 'checked_with_conllu_may.csv'

# news = news[['entities', 'sentenized', 'filt_kw_names', 'language', 'link', 'checked_with_conllu']]

%%time
start = 0
end = len(news)
step = 200

for k in range(start, end, step):
    part = news.iloc[k:k+step].apply(lambda row: sentence_check(row.entities, row.sentenized, row.filt_kw_names, row.language), axis=1)

    news['checked_with_conllu'].update(part)
    news['checked_with_conllu'].to_csv(out_file)
    print(k, news['checked_with_conllu'].notna().sum())
    
    del part

news[['link', 'checked_with_conllu']].to_csv('conllu_checked_may.csv', index=False)

# s = "Головна економістка інвестиційної компанії Dragon CapitaІ Олена Білан сказала виданню, що у грудні уряду необхідно буде залучити приблизно $3 млрд, щоб профінансувати дефіцит бюджету, не урізавши критичних витрат."
# doc = nlp_uk(s)
# conll = doc._.conll_pd
# print(conll)
