==> 6_Searching_for_names_in_dicts.ipynb <==
### 6. Search for filtered names in the dictionary with all names and name forms

import pandas as pd
from uk_stemmer import UkStemmer
from nltk import stem
import re
from collections import Counter
import json

stemmer_ru = stem.snowball.SnowballStemmer("russian") 
stemmer_ukr = UkStemmer()

### Searching for names

entities_path = '../data/entities_may.csv'
entities = pd.read_csv(entities_path, index_col=[0])
entities.columns

# file with all names and name forms gathered during previous months
names_dict_path = 'joined_names_dict.csv'
names_df = pd.read_csv(names_dict_path, index_col=[0])
names_df.head()

# import ast
# entities['names_from_dict'] = entities.names_from_dict.apply(lambda x: ast.literal_eval(x) if pd.notna(x) else None)

entities['unique_kw_names'] = entities.kw_names_unique_str.str.split('§')
entities['set_names'] = entities.one_name_per_set_str.str.split('<@>')

def kw_name_str_to_list(names_and_kw_str):
    if pd.notna(names_and_kw_str):
        res = []
        for ent in names_and_kw_str.split('@+@'):
            res.append(ent.split('<+>'))
        return res
    return None

def stem_names(names_list, stemmer):
    if isinstance(names_list, list): 
        res=[]
        for name in names_list:
            stemmed_name = []
            for word in name.split():
                st = stemmer(word)
                if word[0].isupper():
                    st = st.capitalize()
                stemmed_name.append(st)
            res.append(' '.join(stemmed_name))
        return res
    return None

entities['names_from_dict'] = None
entities['found'] = False

top_names = names_df[(names_df.num>0)]
top_names.shape

def find_names_in_dict(stem_names, found_names, found_flag):
    if isinstance(stem_names, list) and not found_flag:
        res = {}
        if pd.notna(found_names):
            res = found_names
        for i in range(len(stem_names)):
            if (i not in res.keys()) and (stem_names[i] in top_names.index.values):
                res[i] = stem_names[i]
        if len(res) > 0:
            return res
    return None

%%time
part = entities.apply(lambda row: find_names_in_dict(row.set_names, row.names_from_dict, row.found), axis=1)
entities['names_from_dict'].update(part)

entities['found'] = entities.apply(lambda row: 
                           True if (pd.notna(row.names_from_dict) and 
                           len(row.names_from_dict)==len(row.set_names)) else False, axis=1)

entities.found.sum()

not_found=[]
variants=[]

top_names = names_df[names_df.all_vars_str.notna()]
top_names.shape

def find_names_in_variants(found_names, names, found):
    if isinstance(names, list) and not found:
        res = {}
        if pd.notna(found_names):
            res = found_names
        for i in range(len(names)):
            nf = names[i]
            if (i not in res.keys()) and \
                    (nf not in not_found) and \
                        (nf not in variants):
                nf = re.escape(nf)
                pat = rf'§{nf}§|^{nf}§|§{nf}$|^{nf}$'
                found_part = top_names[top_names.all_vars_str.str.contains(pat, na=False)]
                if len(found_part) == 1:
                    res[i] = found_part['stem_name'].iloc[0]
                elif len(found_part) > 1:
                    variants.append(names[i])
                else:
                    not_found.append(names[i])
                    
                del pat, found_part
        if len(res) > 0:
            return res
    return None

%%time
part = entities.apply(lambda row: find_names_in_variants(row.names_from_dict, row.set_names, row.found), axis=1)

entities['names_from_dict'].update(part)


entities[(entities.found==False)&entities.unique_kw_names.notna()].shape

missing = []
def count_missing(stem_names, found_names, found_flag, missing):
    if isinstance(stem_names, list) and not found_flag:
        found={}
        res = [] 
        if pd.notna(found_names):
            found = found_names
        for i in range(len(stem_names)):
            if i not in found.keys():
                res.append(stem_names[i])
        res = list(set(res))
        missing += res
        

entities.apply(lambda row: count_missing(row.unique_kw_names, row.names_from_dict, row.found, missing), axis=1)

missing_counts = Counter(missing)
missing_counts = {k: v for k, v in sorted(missing_counts.items(), key=lambda item: item[1], reverse=True)}
missing_counts

### adding missing names to names dict

# missing_names = [name for name if missing_counts.keys() if re.search(r'[А-ЯЇҐЄІ]\S* [А-ЯЇҐЄІ]\S*', name)]
# for name in missing_names:
#     names_df.loc[name, 'num'] = missing_counts[name]
#     names_df.loc[name, 'id'] = names_df.id.max()+1
# names_df.to_csv(names_dict_path)

entities['names_found'] = entities.names_from_dict.apply(lambda d:
                                                 list(dict(sorted(d.items())).values()) if pd.notna(d) else None)

entities['found_names_str'] = entities.names_found.apply(lambda names: 
                                                    '§'.join(names) if isinstance(names, list) else None)

entities.found_names_str.notna().sum()

entities.columns

entities[['link', 'language', 'all_ent_str', 'name_and_kw', 'kw_and_ent',
       'names_and_kw_str', 'names_str', 'names_sets_str',
       'one_name_per_set_str', 'string_names_sets_str', 'filt_kw_names_str',
       'checked_with_conllu', 'checked_names_str', 'kw_names_unique_str',
       'found', 'found_names_str', 'dict_str']].to_csv(entities_path)

news_filepath = '../data/may.csv'
news = pd.read_csv(news_filepath, index_col=[0])
news.columns

news['found_names_str'] = entities['found_names_str']

news.to_csv(news_filepath)

def dict_to_str(names_from_dict):
    if pd.notna(names_from_dict):
        res = []
        for k, v in names_from_dict.items():
            res.append('§'.join([str(k), v]))
        return '<+>'.join(res)
    return None

entities[entities.names_from_dict.notna()].shape

entities['dict_str'] = entities.names_from_dict.apply(dict_to_str)
