==> 7_Processing_Name_Counts_Dicts.ipynb <==
### 7_Dealing_with_names_counts_dictionaries

##### After counting checked unique name mentions and saving this counts to dataframe, we join different name forms and save these name variations to use them for searching later

import pandas as pd
import re
import json
from collections import Counter

#### Functions for find name variations

names_dict_path = 'joined_names_dict.csv'
names_counts_df = pd.read_csv(names_dict_path, index_col=[0])
names_counts_df.columns

names_counts_df['all_vars'] = names_counts_df.all_vars_str.str.split('§')

def find_lang_variants(name, used_names, ending = '\w{,2}', min_len = (1,3)):
    
    name_var_split = name.split()
    if len(name_var_split)==2:
        if len(name_var_split[0])>min_len[0] and len(name_var_split[1])>min_len[1]:
            name_var = re.sub(r'Во?л(о|а)д', 'Во?л(о|а)д', name_var_split[0])
            name_var = re.sub(r'Паве?л', 'Паве?л', name_var)
            name_var = re.sub(r'Фед(і|о)р', 'Фед(і|о)р', name_var)
            name_var = re.sub(r'В\'?яч', 'В\'?яч', name_var)
            name_var = re.sub(r'Генн?ад', 'Генн?ад', name_var)
            name_var = re.sub(r'Кон?ст(я|а)нт', 'Кон?ст(я|а)нт', name_var)
            name_var = re.sub(r'(Е|О)лен', '(Е|О)лен', name_var)
            name_var = re.sub(r'(М|Н)ик', '(М|Н)ик', name_var)
            name_var = re.sub(r'(О|А)л(е|є|э)кс', '(О|А)л(е|є|э)кс', name_var)
            name_var = re.sub(r'(Х|К)рист', '(Х|К)рист', name_var)
            name_var = re.sub(r'Св(і|е)тл', 'Св(і|е)тл', name_var)
            name_var = re.sub(r'Е?(К|к)атер', 'Е?(К|к)атер', name_var)
            name_var = re.sub(r'Уль?я', 'Уль?я', name_var)

            if name_var != name_var_split[0]:
                upd_part = name_var_split[1]
            else:
                upd_part = name

            upd_part = re.sub(r'[ІИЇЙ]', '(І|И|Ї|Й)', upd_part)
            upd_part = re.sub(r'[ГХҐ]', '(Г|Х|Ґ)', upd_part)
            upd_part = re.sub(r'[хгґ]', '(х|г|ґ)', upd_part)
            upd_part = re.sub(r'[іиїйы]', '(і|и|ї|й|ы)', upd_part)
            upd_part = re.sub(r'[ЕЄЭ]', '(Є|Е|Э)', upd_part)
            upd_part = re.sub(r'[еєэ]', '(е|є|э)', upd_part)
            upd_part = re.sub(r'сь?к', 'сь?к', upd_part)
            upd_part = re.sub(r'ць?к', 'ць?к', upd_part)

            if name_var != name_var_split[0]:
                upd_part = name_var + '?\w{,1} ' + upd_part + ending
            else:
                new_split = upd_part.split()
                upd_part = new_split[0] + '?\w{,1} ' + new_split[1] + ending

            used_names.append(name)
           
            try:
                variants = list(names_counts_df[(~names_counts_df.name.isin(used_names))&(names_counts_df.name.str.contains(rf'^{upd_part}$'))].name.values)
            except Exception as e:
                print(e)
                print(name, ':', upd_part)
                return None

            if len(variants) > 0:
                used_names += variants
                return variants
    return None

used_names = []
lang_vars = names_counts_df[names_counts_df.num>3].name.apply(lambda x: find_lang_variants(x, used_names))

lang_vars[lang_vars.notna()]

def find_contains_variants(name, used_names, min_len = (1,3)):
    name_var_split = name.split()
    if len(name_var_split)==2:
        if len(name_var_split[0])>min_len[0] and len(name_var_split[1])>min_len[1]:
            used_names.append(name)
            try:
                variants = list(names_counts_df[(~names_counts_df.name.isin(used_names))&(names_counts_df.name.str.contains(name))].name.values)
            except Exception as e:
                print(e)
                print(name)
                return None
            
            if len(variants) > 0:
                used_names += variants
                return variants
    return None

used_names2 = []
cont_vars = names_counts_df[names_counts_df.num>3].name.apply(lambda x: find_contains_variants(x, used_names2))

cont_vars[cont_vars.notna()]

names_counts_df['vars1'] = None
names_counts_df['vars2'] = None

names_counts_df.vars1.update(lang_vars)
names_counts_df.vars2.update(cont_vars)

names_counts_df = names_counts_df.set_index('name', drop=False)

def add_same_names_counts(names_counts_df, variants):
    res_df = pd.DataFrame(columns=list(names_counts_df.columns)+['adj_names'])
    for i, r in names_counts_df[names_counts_df[variants].notna()].iterrows():
        names_list = [r.name] + r[variants]
        for sec_name in r[variants]:
            try:
                sec_vars = names_counts_df.loc[sec_name, variants]
            except:
                print(i, sec_name)
                continue   
            if isinstance(sec_vars, list):
                names_list += sec_vars

        res_df.loc[r.name] = names_counts_df[names_counts_df.name.isin(names_list)].sum()
        names_list.remove(r.name)
        res_df.loc[r.name, 'adj_names'] = '§'.join(names_list)
    return res_df

sum_df = add_same_names_counts(names_counts_df, 'vars2')
sum_df['adj_list'] = sum_df.adj_names.str.split('§')

backup = names_counts_df.copy()

# names_counts_df = backup.copy()

names_counts_df.shape

variants = 'vars2'
deleted = []
for i, r in names_counts_df[names_counts_df[variants].notna()].iterrows():
    if r.name not in deleted:
        adj = sum_df.loc[r.name, 'adj_list']
        names_counts_df = names_counts_df[~names_counts_df.name.isin(adj)] 
        deleted += adj
        names_counts_df.loc[r.name, 'num'] = sum_df.loc[r.name, 'num']
        for adj_name in sum_df.loc[r.name, 'adj_list']:
            names_counts_df.loc[r.name, variants].append(adj_name)

def join_two_vars(var1, var2):
    if isinstance(var1, list) and isinstance(var2, list):
        return list(set(var1+var2))
    elif isinstance(var1, list):
        return list(set(var1))
    elif isinstance(var2, list):
        return list(set(var2))
    else:
        return None

names_counts_df['vars3'] = names_counts_df.apply(lambda row: join_two_vars(row.vars1, row.vars2), axis=1)
names_counts_df['all_vars_joined'] = names_counts_df.apply(lambda row: two_vars(row.join_two_vars, row.vars3), axis=1)

names_counts_df.columns

names_counts_df['all_vars'] = names_counts_df['all_vars_joined']
names_counts_df['all_vars_str'] = names_counts_df.all_vars.str.join('§')

names_counts_df.columns

names_counts_df[['name', 'num', 'stem_name', 'all_vars_str', 'id']].to_csv(names_dict_path)

s = pd.Series(index=names_counts_df.index)
s = s.apply(lambda x: [] if x!=x else [])

names_counts_df['all_vars'] = names_counts_df['vars1'].fillna(s) + names_counts_df['vars2'].fillna(s)

names_counts_df['all_vars'] = names_counts_df.all_vars_str.str.split('§')

names_counts_df['all_vars'] = names_counts_df['all_vars'].apply(lambda x: list(set(x)) if isinstance(x,list) else None)

flat_list = [item for sublist in list(names_counts_df[names_counts_df.all_vars.notna()].all_vars.values) for item in sublist]
len(flat_list)

names_counts_df[names_counts_df.index.isin(flat_list)]

# names_counts_df = names_counts_df[~names_counts_df.index.isin(flat_list)]

names_counts_df = clean_up(names_counts_df)

#### Joining dicts

def check_presence(name):
    nf = re.escape(name)
    pat = rf'§{nf}§|^{nf}§|§{nf}$|^{nf}$'
    res = part2[part2.all_vars_str.str.contains(pat)]
    return len(res)

# q = winter_names_counts_df.stem_name.apply(check_presence)

def clean_up(joined_df):
    joined_df['all_vars_str'] = joined_df.all_vars.str.join('§')
    joined_df['all_vars'] = joined_df.all_vars_str.str.split('§')
    joined_df['all_vars'] = joined_df['all_vars'].apply(lambda x: list(set(x)) if isinstance(x,list) else None)
    joined_df['all_vars_str'] = joined_df.all_vars.str.join('§')
    return joined_df

#### Fix names in dicts

def fix_name(joined_df, to_del, to_expand, re_flag):
    if not re_flag:
        try:
            to_del_part = joined_df.loc[to_del]
        except KeyError:
            print(to_del, 'was already fixed')
            return joined_df
        try:
            joined_df.loc[to_expand, 'all_vars'].append(to_del)
            joined_df['all_vars_str'] = joined_df.all_vars.str.join('§')
        except AttributeError:
            joined_df.loc[to_expand, 'all_vars_str'] = to_del
            joined_df['all_vars'] = joined_df.all_vars_str.str.split('§')
            
        joined_df.loc[to_expand, 'num'] += to_del_part.num
        if isinstance(to_del_part.all_vars, list):
            for n in to_del_part.all_vars:
                joined_df.loc[to_expand, 'all_vars'].append(n)
            joined_df['all_vars_str'] = joined_df.all_vars.str.join('§')
        joined_df = joined_df[joined_df.index!=to_del]
                
    else:
        to_del_part = joined_df[(joined_df.index!=to_expand)&joined_df.index.str.contains(to_del)].copy()
        for i,r in to_del_part.iterrows():
            joined_df.loc[to_expand, 'all_vars'].append(i)
            joined_df['all_vars_str'] = joined_df.all_vars.str.join('§')
            joined_df.loc[to_expand, 'num'] += r.num
            if isinstance(r.all_vars, list):
                for n in r.all_vars:
                    joined_df.loc[to_expand, 'all_vars'].append(n)
                joined_df['all_vars_str'] = joined_df.all_vars.str.join('§')
        joined_df = joined_df[~joined_df.index.isin(to_del_part.index.values)]
        del to_del_part
    
    return joined_df

names_counts_df.loc['Королев Єлизавет']

fix_dict = json.load(open('fix_dict.json'))

names_counts_df = fix_name(names_counts_df, ': Шмигал', 'Біньямін Нетаньях', False)

for k,v in fix_dict.items():
    print(names_counts_df.shape)
    names_counts_df = fix_name(names_counts_df, k, v, False)
    print(names_counts_df.shape)

names_counts_df = clean_up(names_counts_df)
names_counts_df[['num', 'stem_name', 'all_vars_str', 'id']].to_csv(names_dict_path)

# to_del = 'Максим Бужанск'
# to_expand = 'Макс Бужанськ'
# to_del_part = joined_df.loc[to_del]
# joined_df.loc[to_expand, 'all_vars'].append(to_del)
# joined_df.loc[to_expand, 'num'] += to_del_part.num
# if isinstance(to_del_part.all_vars, list):
#     for n in to_del_part.all_vars:
#         joined_df.loc[to_expand, 'all_vars'].append(n)
# joined_df = joined_df[joined_df.index!=to_del]
