==> 9_Get_experts_spheres.ipynb <==
### 9. Count experts in every sphere for every category

import pandas as pd
import os
import re
from categories_names import cat_dict

profiles_path = '/Users/oksana/Dev/Масиви по статтях/datasets/experts_profiles.csv'
profiles = pd.read_csv(profiles_path, sep=';')
profiles = profiles.set_index('expert_id', drop=False)

profiles['сфера'].unique()

month='05'
filepath = '/Users/oksana/Dev/Масиви по статтях/datasets/2021/2021_{}/experts_all_{}_2021.csv'.format(month, month)
experts_all = pd.read_csv(filepath, sep=';', )
experts_all = experts_all.set_index('expert_id', drop=False)
# experts_all['num'] = experts_all[experts_all.columns[3:]].sum(axis=1)
experts_all = experts_all[experts_all.columns[:3]]
experts_all['sphere'] = profiles['сфера']

experts_all.sphere = experts_all.sphere.str.strip()

experts_all.head()

experts_all.groupby('sphere').size()

path = '/Users/oksana/Dev/Масиви по статтях/datasets/2021/2021_{}/'.format(month)
datasets = os.listdir(path)
names = []
dfs = {}
for d in datasets:
    if re.match(r'experts', d):
        print(d)
        df = pd.read_csv(path+d, sep=';')
        cat = d.split('_')[1]
        df = df.set_index('expert_id', drop=False)
#         df['num'] = df[df.columns[3:]].sum(axis=1)
        df = df[df.columns[:3]]
        df['sphere'] = profiles['сфера']
        try:
            cat_name = list(cat_dict.keys())[list(cat_dict.values()).index(cat)]
            dfs[cat_name] = df.groupby('sphere').size()
        except:
            pass

dfs.keys()

dfs['Політика']

res = pd.DataFrame(dfs)

res = res.T
res = res.astype('Int64')
res

res.to_csv('sphere_all_{}_2021.csv'.format(month), sep=';')
